\name{tryGetStreams}
\alias{tryGetStreams}
\title{
Get several streams of numbers as columns of a numeric matrix.
}
\description{
The function provides several streams of numbers exactly as generated by
a typical C code application formatted as columns of a numeric matrix.
The seed for the streams is derived from the R base generator.
}
\usage{
tryGetStreams(n, m)
}
\arguments{
  \item{n}{ The length of individual streams.}
  \item{m}{ The number of streams.}
}
\details{
The function is meant for debugging. It provides access from the R level to the
same streams of numbers as which are used in a C code, if stream \code{i} is
initialized with C function \code{initializeRngInstance(k, i)} or the whole set
of streams is created with the C call \code{createRngArray(m, 1)}.
}
\value{
  Numeric matrix with \code{dim=c(n, m)}.
}
\author{
Petr savicky
}
\examples{
set.seed(12345)
tryGetStreams(10, 5)
}

